o1day_mini-rag-service/requirements.txt
fastapi
uvicorn
python-dotenv
pytest
openai
black